[Issue nic_add] Seperate the network code out of vm resource file
